// Backend/src/types/Products/Products.types.d.ts
export type ProductsTypes = {
    id: string;
    codigo: string;
    nome: string;
    categoria: string;
    sku: string;
    preco_venda: number;
    custo: number;
    estoque: number;
    unidade: string;
    descricao: string;
    imagem: string;
    ativo: boolean;
    created_at: string;
    update_at: string;
};
//# sourceMappingURL=Products.types.d.ts.map